backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
openpyxl>=3.1,<3.2
python-dotenv>=1.0,<1.1
